Allow cvdremote to be configured by a file
The file location is provided via an environment variable. That variable
will be defined in a platform-dependent manner when the binary is
distributed (for example, in a wrapper bash script as part of a debian
package)
The file format is TOML (https://toml.io)
Default values for flags are no longer in the code and instead are
expected to be provided through the configuration file.