Check that EnvConfig is convertible to proto.
Since the EnvConfig is converted to proto on the other end, check for
errors by converting to proto on this end. Since we exported the proto
in google/android-cuttlefish#1358, we can import it directly to use it.

(This commit re-lands #437 after revert in #443, ensuring it builds
properly on a bookworm container running go 1.19.)

bug=b/433332364