Avoid copy/paste errors with new cvdr commands
Separate the definition of each command into their own functions.

It's very common to just copy the implementation of the last command and
change what's necessary when adding a new command. It's way to easy to forget
to replace something and use a variable (like the flags) intended for the
existing command in the setup for the new command when those two blocks of code
share the same scope.